// src/app/modules/result/models/assign-subject-list.ts
export class AssignSubjectList {
    CAMPUS_CODE: string;
    USER_CODE : string;
    ORG_CODE: string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
   // SHIFT_CODE: string;
     SECTION_CODE: string;
}



export class ResultAssignSubjectParams {
  ORG_CODE: string;
  CAMPUS_CODE: string;
  CLASS_CODE: string;
  GROUP_CODE: string;
  VERSION_CODE: string;
  SESSION_CODE: string;
  YEAR_CODE: string;
  SEMESTER_CODE: string;
  SHIFT_CODE: string;
  SECTION_CODE: string;
}